Send charts to recycle bin on chart deletion if possible

+++ Quaver.Shared/Database/Maps/MapManager.cs
@@ -9,6 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using Emik;
 using Microsoft.Xna.Framework.Graphics;
 using Quaver.API.Enums;
 using Quaver.API.Maps;
@@ -19,7 +20,9 @@ using Quaver.Shared.Assets;
 using Quaver.Shared.Audio;
 using Quaver.Shared.Config;
 using Quaver.Shared.Database.Playlists;
+using Quaver.Shared.Graphics;
 using Quaver.Shared.Graphics.Backgrounds;
+using Quaver.Shared.Graphics.Dialogs;
 using Quaver.Shared.Graphics.Notifications;
 using Quaver.Shared.Helpers;
 using Quaver.Shared.Modifiers;
@@ -29,6 +32,7 @@ using RestSharp;
 using RestSharp.Extensions;
 using Wobble.Audio.Tracks;
 using Wobble.Bindables;
+using Wobble.Graphics.UI.Dialogs;
 using Wobble.Logging;
 
 namespace Quaver.Shared.Database.Maps
@@ -322,8 +326,29 @@ namespace Quaver.Shared.Database.Maps
                 var mapsetPath = Path.Combine(ConfigManager.SongDirectory.Value, map.Mapset.Directory);
                 var path = Path.Combine(mapsetPath, map.Path);
 
-                File.Delete(path);
-                MapDatabaseCache.RemoveMap(map);
+                if (Rubbish.Move(path))
+                    MapDatabaseCache.RemoveMap(map);
+                else
+                {
+                    ShowFallbackMapDeletionDialog(
+                        "map",
+                        () =>
+                        {
+                            File.Delete(path);
+                            MapDatabaseCache.RemoveMap(map);
+                            map.Mapset.Maps.Remove(map);
+
+                            if (map.Mapset.Maps.Count == 0)
+                                Mapsets.Remove(map.Mapset);
+
+                            PlaylistManager.RemoveMapFromAllPlaylists(map);
+
+                            MapDeleted?.Invoke(typeof(MapManager), new MapDeletedEventArgs(map, index));
+                        }
+                    );
+
+                    return;
+                }
             }
             catch (Exception e)
             {
@@ -372,7 +397,43 @@ namespace Quaver.Shared.Database.Maps
             {
                 var directory = Path.Combine(ConfigManager.SongDirectory.Value, mapset.Directory);
 
-                Directory.Delete(directory, true);
+                if (!Rubbish.Move(directory) && Directory.Exists(directory))
+                {
+                    ShowFallbackMapDeletionDialog(
+                    	"mapset",
+                    	() =>
+	                    {
+                            Directory.Delete(directory, true);
+
+                            try
+                            {
+                                mapset.Maps.ForEach(MapDatabaseCache.RemoveMap);
+                            }
+                            catch (Exception e)
+                            {
+                                Logger.Error(e, LogType.Runtime);
+                            }
+
+                            Mapsets.Remove(mapset);
+                            MapsetDeleted?.Invoke(typeof(MapManager), new MapsetDeletedEventArgs(mapset, index));
+
+                            lock (BackgroundHelper.MapsetBanners)
+                            {
+                                if (!BackgroundHelper.MapsetBanners.ContainsKey(mapset.Directory))
+                                    return;
+
+                                var banner = BackgroundHelper.MapsetBanners[mapset.Directory];
+
+                                if (banner != UserInterface.DefaultBanner)
+                                    banner.Dispose();
+
+                                BackgroundHelper.MapsetBanners.Remove(mapset.Directory);
+                            }
+                        }
+                    );
+
+                    return;
+                }
             }
             catch (Exception e)
             {
@@ -493,5 +554,14 @@ namespace Quaver.Shared.Database.Maps
         {
             SongRequestPlayed?.Invoke(typeof(MapManager), new SongRequestPlayedEventArgs(request, map));
         }
+
+        private static void ShowFallbackMapDeletionDialog(string label, Action onYes) =>
+            DialogManager.Show(
+                new YesNoDialog(
+                    "Map Deletion",
+                    $"Failed to move the {label} in the recycle bin.\nWould you like to delete it instead?",
+                    onYes
+                )
+            );
     }
 }
